updates to implement skeleton pattern
