+++ package.json
@@ -10,14 +10,17 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-native-masked-view/masked-view": "^0.2.8",
     "@react-navigation/native": "^6.1.4",
     "@react-navigation/native-stack": "^6.9.10",
     "axios": "^1.3.3",
     "react": "18.2.0",
     "react-native": "0.71.3",
     "react-native-ios-kit": "^0.0.22",
+    "react-native-linear-gradient": "^2.6.2",
     "react-native-safe-area-context": "^4.5.0",
     "react-native-screens": "^3.20.0",
+    "react-native-skeleton-placeholder": "^5.2.4",
     "react-native-vector-icons": "^9.2.0"
   },
   "devDependencies": {
