+++ src/screens/home/EmptyComponent.js
@@ -1,9 +1,9 @@
-import { Headline } from 'react-native-ios-kit';
+import { Body } from 'react-native-ios-kit';
 import React from 'react';
 
 class EmptyComponent extends React.PureComponent {
     render() {
-        return (< Headline > Add a new note using the 'Add Note' button below</Headline >);
+        return (< Body > No notes found</Body >);
     }
 }
